total active tasks

+++ server/controllers/taskController.js
@@ -6,9 +6,13 @@ const taskController = {};
 taskController.create = async (projectID, taskData) => 
 {
   const newTask = new Task(taskData);
-
   await newTask.save();
-  await Project.updateOne({ id: projectID }, { $push: { tasks: newTask.id } });
+
+  if (taskData.type === 'done')
+    await Project.updateOne({ id: projectID }, { $push: { tasks: newTask.id } });
+
+  else
+    await Project.updateOne({ id: projectID }, { $push: { tasks: newTask.id }, $inc: { activeTasks: 1 } });
 
   console.log(`${new Date()}: successfully inserted task to project |${projectID}|`)
 }
@@ -24,52 +28,67 @@ taskController.updateContent = async (taskID, newContent) =>
 taskController.updateType = async (projectID, taskID, locations, positions) => 
 {
   const project = await Project.findOne({ id: projectID }).lean().select('tasks -_id');
+  const taskList = await Task.find({ id: { $in: project.tasks } });
 
-  const taskIDs = project.tasks;
-  const taskList = await Task.find({ id: { $in: taskIDs } });
-  const taskIndex = taskList.findIndex(task => task.id === taskID);
+  if (locations.old === 'done')
+    await Project.updateOne({ id: projectID }, { $inc: { activeTasks: 1 } });
   
-  await Task.updateOne({ id: taskList[taskIndex].id }, { type: locations.new, position: positions.new, updated_at: new Date() });
+  else if (locations.new === 'done')
+    await Project.updateOne({ id: projectID }, { $inc: { activeTasks: -1 } });
 
-  for (let i = 0; i < taskList.length ; i++) 
-    if (taskList[i].type === locations.old && taskList[i].position > positions.old)
-      await Task.updateOne({ id: taskList[i].id }, { $inc: { position: -1 } });
+  await Promise.all(taskList.map(async task => 
+  {
+    if (task.type === locations.old && task.position > positions.old) 
+      await Task.updateOne({ id: task.id }, { $inc: { position: -1 } });
+
+    else if (task.id === taskID) 
+      await Task.updateOne({ id: taskID }, { type: locations.new, position: positions.new, updated_at: new Date() });
+  }));
   
   console.log(`${new Date()}: successfully moved task to |${locations.new}|`);
 }
 
+/*********************************************************************************************************************************/
 taskController.updatePosition = async (updatedTaskID, otherTaskID, direction) =>
 {
-  if (direction === 'up')
+  switch (direction)
   {
-    await Task.updateOne({ id: updatedTaskID }, { $inc: { position: -1 } });
-    await Task.updateOne({ id: otherTaskID }, { $inc: { position: +1 } });
+    case 'up':
+      await Task.updateOne({ id: updatedTaskID }, { $inc: { position: -1 } });
+      await Task.updateOne({ id: otherTaskID }, { $inc: { position: +1 } });
+      break;
+
+    case 'down':
+      await Task.updateOne({ id: updatedTaskID }, { $inc: { position: +1 } });
+      await Task.updateOne({ id: otherTaskID }, { $inc: { position: -1 } });
+      break;
   }
 
-  else if (direction === 'down')
-  {
-    await Task.updateOne({ id: updatedTaskID }, { $inc: { position: +1 } });
-    await Task.updateOne({ id: otherTaskID }, { $inc: { position: -1 } });
-  }
+  console.log(`${new Date()}: successfully updated task position`);
 }
 
 /*********************************************************************************************************************************/
 taskController.delete = async (projectID, taskID, taskType, position) =>
 {
   const project = await Project.findOne({ id: projectID }).lean().select('tasks -_id');
+  project.tasks.splice(project.tasks.findIndex(task => task === taskID), 1);
   
-  const taskIDs = project.tasks;
-  const taskList = await Task.find({ id: { $in: taskIDs } });
-  const taskIndex = taskIDs.findIndex(task => task === taskID);
-  taskIDs.splice(taskIndex, 1);
+  const taskList = await Task.find({ id: { $in: project.tasks } });
 
-  await Project.updateOne({ id: projectID }, { $set: { tasks: taskIDs } });
-  await Task.deleteOne({ id: taskID });
+  if (taskType === 'done')
+    await Project.updateOne({ id: projectID }, { $pull: { tasks: taskID } });
+  
+  else
+    await Project.updateOne({ id: projectID }, { $pull: { tasks: taskID }, $inc: { activeTasks: -1 } });
 
-  for (let i = 0; i < taskList.length ; i++) 
-    if (taskList[i].type === taskType && taskList[i].position > position)
-      await Task.updateOne({ id: taskList[i].id }, { $inc: { position: -1 } });
-      console.log('updated task position');
+  await Promise.all(taskList.map(async task => 
+  {
+    if (task.type === taskType && task.position > position) 
+      await Task.updateOne({ id: task.id }, { $inc: { position: -1 } });
+      
+    else if (task.id === taskID)
+      await Task.deleteOne({ id: taskID });
+  }));
 
   console.log(`${new Date()}: successfully deleted task |${taskID}|`);
 }
